TA web app/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
jinja2
websockets
websocket-client
